requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
motor==3.6.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.17
python-dotenv==1.0.1
pydantic==2.10.2
bcrypt==4.2.1
email-validator==2.2.0